remove coupling to sbt-projectmatrix
See https://github.com/scalacenter/scalafix.g8/pull/19
