+++ project/plugins.sbt
@@ -1,6 +1,5 @@
 resolvers += Resolver.sonatypeRepo("public")
 addSbtPlugin("com.geirsson" % "sbt-ci-release" % "1.5.7")
-addSbtPlugin("com.eed3si9n" % "sbt-buildinfo" % "0.10.0")
 
 libraryDependencies += "org.scala-sbt" %% "scripted-plugin" % sbtVersion.value
 unmanagedSourceDirectories.in(Compile) ++= {
