+++ src/main/scala-sbt-1.0/scalafix/sbt/ScalafixTestkitPlugin.scala
@@ -2,9 +2,7 @@ package scalafix.sbt
 
 import sbt.Keys._
 import sbt._
-import sbt.internal.ProjectMatrix
 import sbt.plugins.JvmPlugin
-import sbtprojectmatrix.ProjectMatrixPlugin.autoImport._
 
 import java.io.File.pathSeparator
 
@@ -13,43 +11,6 @@ object ScalafixTestkitPlugin extends AutoPlugin {
   override def requires: Plugins = JvmPlugin
 
   object autoImport {
-    case class TestkitTargetAxis(scalaVersion: String)
-        extends VirtualAxis.WeakAxis {
-      private val scalaBinaryVersion =
-        CrossVersion.binaryScalaVersion(scalaVersion)
-
-      override val idSuffix = s"Target${scalaBinaryVersion.replace('.', '_')}"
-      override val directorySuffix = s"target$scalaBinaryVersion"
-    }
-
-    object TestkitTargetAxis {
-
-      private def targetScalaVersion(virtualAxes: Seq[VirtualAxis]): String =
-        virtualAxes.collectFirst { case a: TestkitTargetAxis =>
-          a.scalaVersion
-        }.get
-
-      def resolve[T](
-          matrix: ProjectMatrix,
-          key: TaskKey[T]
-      ): Def.Initialize[Task[T]] =
-        Def.taskDyn {
-          val sv = targetScalaVersion(virtualAxes.value)
-          val project = matrix.finder().apply(sv)
-          Def.task((project / key).value)
-        }
-
-      def resolve[T](
-          matrix: ProjectMatrix,
-          key: SettingKey[T]
-      ): Def.Initialize[T] =
-        Def.settingDyn {
-          val sv = targetScalaVersion(virtualAxes.value)
-          val project = matrix.finder().apply(sv)
-          Def.setting((project / key).value)
-        }
-    }
-
     val scalafixTestkitInputClasspath =
       taskKey[Classpath]("Classpath of input project")
     val scalafixTestkitInputScalacOptions =
